requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.6.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
colorlog==6.8.0
schedule==1.2.0
email-validator==2.1.0